requirements.txt
slackclient>=2.0.0
certifi
slackeventsapi>=0.1